+++ leia-common/src/main/java/com/grookage/leia/common/validation/ValidationContext.java
@@ -1,21 +1,22 @@
-package com.grookage.leia.common.violation;
+package com.grookage.leia.common.validation;
 
+import com.grookage.leia.common.violation.LeiaSchemaViolation;
+import com.grookage.leia.common.violation.LeiaSchemaViolationImpl;
 import lombok.Builder;
 import lombok.Data;
 import lombok.Getter;
-import lombok.NoArgsConstructor;
+import lombok.RequiredArgsConstructor;
 
-import java.util.ArrayList;
-import java.util.LinkedList;
-import java.util.List;
+import java.util.*;
 
 @Data
 @Builder
-@NoArgsConstructor
-public class ViolationContext {
+@RequiredArgsConstructor
+public class ValidationContext {
     @Getter
     private final List<LeiaSchemaViolation> violations = new ArrayList<>();
     private final LinkedList<Class<?>> klassPath = new LinkedList<>();
+    private final Set<Class<?>> validationQueue = new HashSet<>();
 
     public void addViolation(final String message) {
         violations.add(new LeiaSchemaViolationImpl(message, null, klassPath.peekLast()));
@@ -26,6 +27,10 @@ public class ViolationContext {
         violations.add(new LeiaSchemaViolationImpl(message, path, klassPath.peekLast()));
     }
 
+    public void addViolation(List<LeiaSchemaViolation> schemaViolations) {
+        violations.addAll(schemaViolations);
+    }
+
     public void pushClass(final Class<?> klass) {
         klassPath.addLast(klass);
     }
@@ -34,5 +39,9 @@ public class ViolationContext {
         if (!klassPath.isEmpty()) {
             klassPath.removeLast();
         }
+    }
+
+    public void validateClass(final Class<?> klass) {
+        validationQueue.add(klass);
     }
 }